GO-5341 Resolve layout via details svc
