+++ core/block/editor/template/template.go
@@ -339,6 +339,47 @@ var WithNameToFirstBlock = StateTransformer(func(s *state.State) {
 	}
 })
 
+var WithNameFromFirstBlock = StateTransformer(func(s *state.State) {
+	name, ok := s.Details().TryString(bundle.RelationKeyName)
+
+	if !ok || name == "" {
+		textBlock, err := getFirstTextBlock(s)
+		if err != nil {
+			log.Errorf("failed to get first block with text: %v", err)
+			return
+		}
+		if textBlock == nil {
+			return
+		}
+		s.SetDetail(bundle.RelationKeyName, domain.String(textBlock.Model().GetText().GetText()))
+
+		for _, id := range textBlock.Model().ChildrenIds {
+			s.Unlink(id)
+		}
+		err = s.InsertTo(textBlock.Model().Id, model.Block_Bottom, textBlock.Model().ChildrenIds...)
+		if err != nil {
+			log.Errorf("insert children: %v", err)
+			return
+		}
+		s.Unlink(textBlock.Model().Id)
+	}
+})
+
+func getFirstTextBlock(st *state.State) (simple.Block, error) {
+	var res simple.Block
+	err := st.Iterate(func(b simple.Block) (isContinue bool) {
+		if b.Model().GetText() != nil {
+			res = b
+			return false
+		}
+		return true
+	})
+	if err != nil {
+		return nil, err
+	}
+	return res, nil
+}
+
 var WithFeaturedRelationsBlock = StateTransformer(func(s *state.State) {
 	RequireHeader(s)
