+++ pkg/lib/bundle/relations.json
@@ -1535,5 +1535,15 @@
     "name": "Plural name",
     "readonly": false,
     "source": "details"
+  },
+  {
+    "description": "Layout settings are strictly inherited for all objects of this type",
+    "format": "checkbox",
+    "hidden": true,
+    "key": "forceLayoutFromType",
+    "maxCount": 1,
+    "name": "Force layout from type",
+    "readonly": false,
+    "source": "details"
   }
 ]
